third_party/blink/web_tests/http/tests/cookies/http-get-cookie-set-in-js-expected.txt

Test that a cookie set using JavaScript can be correctly read by HTTP server

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

checking that the cookie set in javascript can be read by HTTP server
PASS cookie is 'name=value'.
PASS successfullyParsed is true

TEST COMPLETE
